Separated WebNavigator and GitHubScraper into different classes.
